# bff/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend For Frontend API",
    "description": "API documentation for the BFF service.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "operationId": "get_all_reviews",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Max number of reviews to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 100
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "List all reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WrapperReview"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "operationId": "create_review",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapperNewReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create a new review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperReview"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/reviews/{review_id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "operationId": "get_review_by_id",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "description": "ID of the review to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get review by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperReview"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "operationId": "update_review",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "description": "ID of the review to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WrapperNewReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update review by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrapperReview"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "operationId": "delete_review",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "description": "ID of the review to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete review by ID",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/{user_id}/watchlist": {
      "get": {
        "tags": [
          "Watchlist"
        ],
        "operationId": "get_watchlist_by_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "UUID of the user whose watchlist to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max number of items to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "List user watchlist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Watchlist"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{user_id}/watchlist/{content_id}": {
      "delete": {
        "tags": [
          "Watchlist"
        ],
        "operationId": "delete_watchlist_item",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "UUID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content_id",
            "in": "path",
            "description": "UUID of the content item",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete item from watchlist",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/watchlist": {
      "post": {
        "tags": [
          "Watchlist"
        ],
        "operationId": "create_watchlist_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewWatchlist"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add a new item to the watchlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watchlist"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "NewWatchlist": {
        "type": "object",
        "required": [
          "user_id",
          "content_id"
        ],
        "properties": {
          "content_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Watchlist": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "content_id",
          "created_at"
        ],
        "properties": {
          "content_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "WrapperNewReview": {
        "type": "object",
        "required": [
          "movie_id",
          "rating",
          "user_name"
        ],
        "properties": {
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "movie_id": {
            "type": "integer",
            "format": "int32"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_name": {
            "type": "string"
          }
        }
      },
      "WrapperReview": {
        "type": "object",
        "required": [
          "created_at",
          "id",
          "movie_id",
          "rating",
          "user_name"
        ],
        "properties": {
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "movie_id": {
            "type": "integer",
            "format": "int32"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_name": {
            "type": "string"
          }
        }
      }
    }
  }
}